#1. 모듈화
- 모듈화 진행
- 모험가 해적 링크 적용
