+++ app/templates/components/_modals.html
@@ -107,12 +107,12 @@
                     <br>
                     <div class="flex gap-4">
                         <div class="w-8 h-8 rounded-full bg-blue-100 text-blue-600 flex items-center justify-center font-bold flex-none">3</div>
-                        <p class="text-sm text-slate-600 leading-relaxed"><span class="font-bold text-slate-800">상세 진단을 활용하세요.</span> 장비 아이콘을 클릭하면 세부 수치와 함께 교체 시점, 강화 권장 사항 등 맞춤형 조언을 확인할 수 있습니다.</p>
+                        <p class="text-sm text-slate-600 leading-relaxed"><span class="font-bold text-slate-800">전체 아이템 진단을 활용하세요.</span> 전체 아이템 진단 아이콘을 클릭하면 모든 아이템의 조언을 확인할 수 있습니다.</p>
                     </div>
                     <br>
                     <div class="flex gap-4">
                         <div class="w-8 h-8 rounded-full bg-blue-100 text-red-600 flex items-center justify-center font-bold flex-none">🚨</div>
-                        <p class="text-sm text-slate-600 leading-relaxed"><span class="font-bold text-slate-800">특수 직업의 경우 출력이 부정확할 수 있습니다.</span> 모험가 해적 링크는 현재 출력되지 않습니다. / 특수 직업의 경우 자체적인 로직으로 인해 점수가 과도하게 높거나 낮게 설정될 수 있습니다.</p>
+                        <p class="text-sm text-slate-600 leading-relaxed"><span class="font-bold text-slate-800">특수 직업의 경우 출력이 부정확할 수 있습니다.</span> 특수 직업의 경우 자체적인 로직으로 인해 점수가 과도하게 높거나 낮게 설정될 수 있습니다.</p>
                     </div>
                 </div>
             </div>
